backend/requirements2.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
coverage==7.3.2
dj-database-url==2.1.0
Django==4.2.5
django-cas-ng==5.0.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
flake8==6.1.0
frozenlist==1.4.0
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.3
mccabe==0.7.0
multidict==6.0.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pluggy==1.3.0
postgres==4.0
promise==2.3
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.6.0
pytest-pythonpath==0.7.3
python-cas==1.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
